requirements.txt
opencv-python
pdf2image
pillow
pyocr
